requirements.txt
Flask==2.3.2
pycryptodome==3.23.0
